requirements.txt
Django~=4.2
environs~=9.5.0
python-telegram-bot~=13.0